# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OOTD API
  version: 1.0.0
  description: |
    Mobile OOTD API Server Specification.
    Authentication uses JWT Bearer tokens issued after Kakao OAuth2 login.
    All API responses are wrapped by a common response object:
    { success, code, message, data }.
servers:
  - url: http://localhost:8080
    description: Local
tags:
  - name: Auth
  - name: Posts
  - name: Engagement
paths:
  /api/auth/kakao/login:
    post:
      tags: [Auth]
      summary: Login or sign up with Kakao OAuth2 authorization code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KakaoLoginRequest'
      responses:
        '201':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Reissue access/refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Reissued tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '401':
          description: Invalid refresh token
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user (delete refresh token)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
        '401':
          description: Unauthorized
  /api/posts:
    post:
      tags: [Posts]
      summary: Create post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Unauthorized
  /api/posts/feed:
    get:
      tags: [Posts]
      summary: Get weather/region filtered feed
      parameters:
        - in: query
          name: weatherType
          schema:
            $ref: '#/components/schemas/WeatherType'
        - in: query
          name: region
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Feed page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePostResponse'
  /api/posts/popular:
    get:
      tags: [Posts]
      summary: Get popular posts by Redis score
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Popular posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PopularPostResponse'
  /api/posts/{postId}/likes:
    post:
      tags: [Engagement]
      summary: Like a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '204':
          description: Liked
        '401':
          description: Unauthorized
        '404':
          description: Post not found
        '409':
          description: Already liked
    delete:
      tags: [Engagement]
      summary: Unlike a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '204':
          description: Unliked
        '401':
          description: Unauthorized
        '404':
          description: Like not found
  /api/posts/{postId}/views:
    post:
      tags: [Engagement]
      summary: Record post view event
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '204':
          description: View recorded
        '404':
          description: Post not found
  /api/posts/{postId}/comments:
    post:
      tags: [Engagement]
      summary: Create comment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/postId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '401':
          description: Unauthorized
        '404':
          description: Post not found
    get:
      tags: [Engagement]
      summary: Get recent comments (up to 50)
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '200':
          description: Recent comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    postId:
      in: path
      name: postId
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    WeatherType:
      type: string
      enum: [SUNNY, CLOUDY, RAINY, SNOWY, WINDY]
    KakaoLoginRequest:
      type: object
      required: [authorizationCode, redirectUri]
      properties:
        authorizationCode:
          type: string
        redirectUri:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        nickname:
          type: string
    AccessTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
    CreatePostRequest:
      type: object
      required: [weatherType, region, content]
      properties:
        weatherType:
          $ref: '#/components/schemas/WeatherType'
        region:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    PostResponse:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        authorId:
          type: integer
          format: int64
        authorNickname:
          type: string
        weatherType:
          $ref: '#/components/schemas/WeatherType'
        region:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        likeCount:
          type: integer
          format: int64
        viewCount:
          type: integer
          format: int64
        commentCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    CommentResponse:
      type: object
      properties:
        commentId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    PopularPostResponse:
      type: object
      properties:
        post:
          $ref: '#/components/schemas/PostResponse'
        score:
          type: number
          format: double
    PagePostResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean